pdf-toolkit/scripts/requirements.txt
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pillow>=10.0.0
reportlab>=4.0.0
